+++ CHANGES.md
@@ -7,6 +7,8 @@ version ?????????
 - continuous or discrete restriction mode for state constraints, ban, fence
 - btrpsl joined the repository
 - back to a new fast Packing constraint thanks to Sophie Demassey
+- ChocoReconfigurationAlgorithmParams renamed to Parameters
+- bug fixes
 
 version 0.38.2 - 19 Jun 2014
 ----------------------